gdelospi - scerrada

+++ src/main/java/etl/rolap/repositorios/PacienteRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<DimPaciente,Integer> {
-
+public interface PacienteRepository extends JpaRepository<DimPaciente,Long> {
 
 }
